Include accumulated turn content in chunk mutation
