+++ .changeset/sour-seahorses-walk.md
@@ -1,0 +1,5 @@
+---
+'@aws-amplify/ai-constructs': minor
+---
+
+Include accumulated turn content in chunk mutation
