+++ packages/ai-constructs/src/conversation/runtime/types.ts
@@ -100,6 +100,7 @@ export type StreamingResponseChunk = {
   conversationId: string;
   associatedUserMessageId: string;
   contentBlockIndex: number;
+  accumulatedTurnContent: Array<bedrock.ContentBlock>;
 } & (
   | {
       // text chunk
